Moved equality comparers back to test
