+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/test_utils/AnyOrderCollectionEqualityComparer.java
@@ -1,7 +1,5 @@
 package com.scottlogic.deg.generator.decisiontree.test_utils;
 
-import com.scottlogic.deg.generator.utils.EqualityComparer;
-
 import java.util.ArrayList;
 import java.util.Collection;
