+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/test_utils/ConstraintNodeComparer.java
@@ -2,7 +2,6 @@ package com.scottlogic.deg.generator.decisiontree.test_utils;
 
 import com.scottlogic.deg.generator.decisiontree.ConstraintNode;
 import com.scottlogic.deg.generator.decisiontree.DecisionNode;
-import com.scottlogic.deg.generator.utils.EqualityComparer;
 
 import java.util.Collection;
