+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/test_utils/DefaultEqualityComparer.java
@@ -1,7 +1,5 @@
 package com.scottlogic.deg.generator.decisiontree.test_utils;
 
-import com.scottlogic.deg.generator.utils.EqualityComparer;
-
 public class DefaultEqualityComparer implements EqualityComparer {
     private final EqualityComparer collectionComparer;
