+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/test_utils/EqualityComparer.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.utils;
+package com.scottlogic.deg.generator.decisiontree.test_utils;
 
 public interface EqualityComparer {
     int getHashCode(Object item);
