+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/test_utils/ProfileFieldComparer.java
@@ -1,7 +1,6 @@
 package com.scottlogic.deg.generator.decisiontree.test_utils;
 
 import com.scottlogic.deg.generator.ProfileFields;
-import com.scottlogic.deg.generator.utils.EqualityComparer;
 
 import java.util.Collection;
 import java.util.stream.Collectors;
