+++ generator/src/test/java/com/scottlogic/deg/generator/restrictions/DateTimeRestrictionsEqualityComparer.java
@@ -1,7 +1,6 @@
-package com.scottlogic.deg.generator.restrictions.Equality;
+package com.scottlogic.deg.generator.restrictions;
 
-import com.scottlogic.deg.generator.restrictions.DateTimeRestrictions;
-import com.scottlogic.deg.generator.utils.EqualityComparer;
+import com.scottlogic.deg.generator.decisiontree.test_utils.EqualityComparer;
 
 public class DateTimeRestrictionsEqualityComparer implements EqualityComparer {
