+++ generator/src/test/java/com/scottlogic/deg/generator/restrictions/FieldToFieldSpecComparer.java
@@ -1,8 +1,7 @@
-package com.scottlogic.deg.generator.restrictions.Equality;
+package com.scottlogic.deg.generator.restrictions;
 
 import com.scottlogic.deg.generator.Field;
-import com.scottlogic.deg.generator.restrictions.FieldSpec;
-import com.scottlogic.deg.generator.utils.EqualityComparer;
+import com.scottlogic.deg.generator.decisiontree.test_utils.EqualityComparer;
 
 import java.util.Map;
